requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
wsgiref==0.1.2
celery-with-redis==3.0
redis==2.7.2
hiredis==0.1.1
requests==1.2.0